refactor: #107 friendUiModel 패키지 변경
