+++ presenter/src/main/java/com/mashup/chinchin/presenter/main/model/RecommendFriendUiModel.kt
@@ -1,5 +1,7 @@
 package com.mashup.chinchin.presenter.main.model
 
+import com.mashup.chinchin.presenter.common.model.FriendUiModel
+
 data class RecommendFriendUiModel(
     val id: Long,
     val profileUrl: String,
