+++ presenter/src/main/java/com/mashup/chinchin/presenter/ui/common/ChinChinComponent.kt
@@ -29,7 +29,7 @@ import com.mashup.chinchin.presenter.R
 import com.mashup.chinchin.presenter.common.CardState
 import com.mashup.chinchin.presenter.common.ChinChinAnswerCardState
 import com.mashup.chinchin.presenter.common.ChinChinQuestionCardState
-import com.mashup.chinchin.presenter.main.model.FriendUiModel
+import com.mashup.chinchin.presenter.common.model.FriendUiModel
 import com.mashup.chinchin.presenter.ui.theme.*
 
 @Composable
