+++ presenter/src/main/java/com/mashup/chinchin/presenter/ui/main/home/HomeComponent.kt
@@ -29,7 +29,7 @@ import coil.compose.AsyncImage
 import com.mashup.chinchin.presenter.R
 import com.mashup.chinchin.presenter.group_detail.GroupDetailActivity
 import com.mashup.chinchin.presenter.main.model.FriendGroupUiModel
-import com.mashup.chinchin.presenter.main.model.FriendUiModel
+import com.mashup.chinchin.presenter.common.model.FriendUiModel
 import com.mashup.chinchin.presenter.ui.common.ChinChinButton
 import com.mashup.chinchin.presenter.ui.common.ChinChinGrayTextField
 import com.mashup.chinchin.presenter.ui.common.ChinChinText
